Started Bomb Squad and added Minor Fixes
Added the Bomb Squad project that will be using the Engine to demonstrate a practicial application of the Objx library to game development as well as added minor fixes to the SDL2_Engine in response to issues that where encountered during initial setup

+++ SDL2_Engine/Projects/SDL2_Engine/src/Input/Controllers/Controllers.cpp
@@ -195,7 +195,7 @@ namespace SDL2_Engine {
 		*/
 		const float Controllers::rawAxis(const EGamePadAxisCodes& pAxis, const Utilities::Bitmask<EGamePadID>& pIDs /*= EGamePadID::All*/) const noexcept { 
 			//Check if the mask is multiple or a single ID
-			int s = fsqrt(pIDs.getMask());
+			int s = (int)fsqrt(pIDs.getMask());
 			if ((s * s) == pIDs.getMask()) 
 				return mData->gamepads[ID2IND(pIDs.getMask())].rawAxis(pAxis);
 
@@ -231,7 +231,7 @@ namespace SDL2_Engine {
 		*/
 		const float Controllers::rawAxisDelta(const EGamePadAxisCodes& pAxis, const Utilities::Bitmask<EGamePadID>& pIDs /*= EGamePadID::All*/) const noexcept { 
 			//Check if the mask is multiple or a single ID
-			int s = fsqrt(pIDs.getMask());
+			int s = (int)fsqrt(pIDs.getMask());
 			if ((s * s) == pIDs.getMask())
 				return mData->gamepads[ID2IND(pIDs.getMask())].rawAxisDelta(pAxis);
